MSI-2882: Store Pickup button present on check-out when products are not not available for pickup. Moved API interface to the model namespace; Moved implementation to the implementation module.
