+++ InventoryInStorePickupShippingApi/Model/Carrier/InStorePickup.php
@@ -108,11 +108,11 @@ class InStorePickup extends AbstractCarrier implements CarrierInterface
         $this->requestValidator = $requestValidator;
         $this->getFreePackages = $getFreePackages;
         $this->getShippingPriceRequest = $getShippingPriceRequest;
+        $this->getCarrierTitle = $getCarrierTitle;
 
         $this->_code = self::CARRIER_CODE;
 
         parent::__construct($scopeConfig, $rateErrorFactory, $logger, $data);
-        $this->getCarrierTitle = $getCarrierTitle;
     }
 
     /**
