+++ InventoryInStorePickupShippingApi/Model/IsInStorePickupDeliveryAvailableForCartInterface.php
@@ -5,7 +5,7 @@
  */
 declare(strict_types=1);
 
-namespace Magento\InventoryInStorePickupShippingApi\Api;
+namespace Magento\InventoryInStorePickupShippingApi\Model;
 
 /**
  * Check if In-Store Pickup delivery method is applicable for a cart by cartId.
